o365-pull-connector/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "o365-connector"
version = "0.1.0"
description = "Standalone Microsoft 365 pull connector"
authors = [{name = "o365-connector"}]
dependencies = [
    "Flask>=3.0.0",
    "APScheduler>=3.10.0",
    "SQLAlchemy>=2.0.0",
    "requests>=2.31.0",
    "msal>=1.26.0",
    "PyYAML>=6.0.1",
]
